Add Knowledge Graph component with graph generation, navigation, and visualization features. Implement user interface elements for generating graphs, displaying navigation controls, and visualizing graph data. Introduce toast notifications for user feedback and a delete confirmation modal for managing graphs. Enhance overall user experience with improved layout and accessibility.
