+++ src/frontend/hooks/useKeyboardShortcuts.tsx
@@ -1,0 +1,30 @@
+import { useEffect } from 'react';
+
+interface KeyboardShortcuts {
+  onPrevious?: () => void;
+  onNext?: () => void;
+  onDelete?: () => void;
+}
+
+export function useKeyboardShortcuts({ onPrevious, onNext, onDelete }: KeyboardShortcuts) {
+  useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      // Don't trigger shortcuts when typing in input fields
+      if (e.target instanceof HTMLInputElement || e.target instanceof HTMLTextAreaElement) return;
+      
+      if (e.key === 'ArrowLeft' && onPrevious) {
+        e.preventDefault();
+        onPrevious();
+      } else if (e.key === 'ArrowRight' && onNext) {
+        e.preventDefault();
+        onNext();
+      } else if ((e.key === 'Delete' || e.key === 'Backspace') && onDelete) {
+        e.preventDefault();
+        onDelete();
+      }
+    };
+
+    window.addEventListener('keydown', handleKeyDown);
+    return () => window.removeEventListener('keydown', handleKeyDown);
+  }, [onPrevious, onNext, onDelete]);
+}
