+++ src/frontend/hooks/useLocalStorage.tsx
@@ -1,0 +1,51 @@
+import { useState } from 'react';
+
+export function useLocalStorage<T>(key: string, defaultValue: T) {
+  const [value, setValue] = useState<T>(() => {
+    if (typeof window === 'undefined') {
+      return defaultValue;
+    }
+    
+    try {
+      const item = localStorage.getItem(key);
+      return item ? JSON.parse(item) : defaultValue;
+    } catch (error) {
+      console.warn(`Failed to load from localStorage for key "${key}":`, error);
+      return defaultValue;
+    }
+  });
+
+  const setStoredValue = (newValue: T | ((val: T) => T)) => {
+    try {
+      const valueToStore = newValue instanceof Function ? newValue(value) : newValue;
+      setValue(valueToStore);
+      
+      if (typeof window !== 'undefined') {
+        localStorage.setItem(key, JSON.stringify(valueToStore));
+      }
+    } catch (error) {
+      console.warn(`Failed to save to localStorage for key "${key}":`, error);
+    }
+  };
+
+  return [value, setStoredValue] as const;
+}
+
+// Utility functions for direct localStorage operations
+export function saveToLocalStorage<T>(key: string, data: T): void {
+  try {
+    localStorage.setItem(key, JSON.stringify(data));
+  } catch (error) {
+    console.warn('Failed to save to localStorage:', error);
+  }
+}
+
+export function loadFromLocalStorage<T>(key: string, defaultValue: T): T {
+  try {
+    const item = localStorage.getItem(key);
+    return item ? JSON.parse(item) : defaultValue;
+  } catch (error) {
+    console.warn('Failed to load from localStorage:', error);
+    return defaultValue;
+  }
+}
